Move file, small code changes
