+++ app/templates/events/edit.html.j2
@@ -9,9 +9,7 @@
 
 {% include "events/edit/_info.html.j2" %}
 
-<div class="mb-3">
-
-</div>
+<div class="mb-3"></div>
 
 <hr>
